Move AnimalCombination
This PR:
- Renames and moves `RUFAS/enums.py` to `RUFAS/biophysical/animal/data_types/animal_combination.py`. The original `enums.py` only houses the `AnimalCombination` enum class, which is only needed by the animal module, hence the renaming and moving.
- Moves `RUFAS/data_structures/animal_manure_excretions.py` to `RUFAS/biophysical/animal/data_types/animal_manure_excretions.py`. The `AnimalManureExcretions` data type is only used in the animal module.
- Removes the old `PenManureData` in `RUFAS/data_structures/pen_manure_data.py`
- Update all the references to the above files and the unit tests.
