+++ RUFAS/biophysical/animal/data_types/animal_combination.py
