+++ RUFAS/biophysical/animal/data_types/animal_manure_excretions.py
