+++ RUFAS/biophysical/animal/digestive_system/manure_excretion_calculator.py
@@ -2,7 +2,7 @@ from typing import Tuple
 
 from RUFAS.biophysical.animal.animal_module_constants import AnimalModuleConstants
 from RUFAS.biophysical.animal.data_types.nutrition_data_structures import NutritionSupply
-from RUFAS.data_structures.animal_manure_excretions import AnimalManureExcretions
+from RUFAS.biophysical.animal.data_types.animal_manure_excretions import AnimalManureExcretions
 
 from RUFAS.general_constants import GeneralConstants
