+++ RUFAS/biophysical/animal/ration/ration_optimizer.py
@@ -5,7 +5,7 @@ import scipy
 import numpy.typing as npt
 from typing import List, Tuple, Dict, Callable, Any
 from RUFAS.biophysical.animal.nutrients.nutrition_supply_calculator import NutritionSupplyCalculator, FeedInRation
-from RUFAS.enums import AnimalCombination
+from RUFAS.biophysical.animal.data_types.animal_combination import AnimalCombination
 from RUFAS.units import MeasurementUnits
 from RUFAS.general_constants import GeneralConstants
 from RUFAS.biophysical.animal.animal_module_constants import AnimalModuleConstants
