+++ RUFAS/data_structures/animal_to_manure_connection.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 from enum import Enum
 from typing import Optional
-from RUFAS.enums import AnimalCombination
+from RUFAS.biophysical.animal.data_types.animal_combination import AnimalCombination
 from RUFAS.output_manager import OutputManager
 from RUFAS.units import MeasurementUnits
