+++ RUFAS/routines/animal/ration/user_defined_ration.py
@@ -1,6 +1,6 @@
 from typing import Dict
 
-from RUFAS.enums import AnimalCombination
+from RUFAS.biophysical.animal.data_types.animal_combination import AnimalCombination
 from RUFAS.routines.animal.animal_typed_dicts import AvailableFeedsTypedDict
 from RUFAS.routines.animal.ration.animal_requirements import AnimalRequirements
