+++ tests/_test_manure_module/_test_manure_manager_constants.py
