+++ tests/_test_manure_module/_test_pen_manure.py
@@ -3,7 +3,7 @@ from dataclasses import fields
 import pytest
 from pytest import approx
 
-from RUFAS.data_structures.animal_manure_excretions import AnimalManureExcretions
+from RUFAS.biophysical.animal.data_types.animal_manure_excretions import AnimalManureExcretions
 from RUFAS.general_constants import GeneralConstants
 from RUFAS.routines.manure.constants_and_units.manure_constants import ManureConstants
 from RUFAS.routines.manure.pen_manure.pen_manure import PenManure
