+++ tests/_test_manure_module/enums/_test_manure_cover_enum.py
