+++ tests/_test_manure_module/test_animal_manure_excretions/_test_general_manure.py
@@ -1,7 +1,7 @@
 import pytest
 from pytest import approx
 
-from RUFAS.data_structures.animal_manure_excretions import AnimalManureExcretions
+from RUFAS.biophysical.animal.data_types.animal_manure_excretions import AnimalManureExcretions
 from RUFAS.general_constants import GeneralConstants
 from RUFAS.routines.animal.manure.general_manure import calculate_phosphorus_excretion_values
