+++ tests/test_biophysical/test_animal/test_herd_manager/test_herd_manager_properties.py
@@ -2,7 +2,7 @@ from RUFAS.biophysical.animal.animal import Animal
 from RUFAS.biophysical.animal.herd_manager import HerdManager
 from RUFAS.biophysical.animal.data_types.animal_types import AnimalType
 from RUFAS.biophysical.animal.pen import Pen
-from RUFAS.enums import AnimalCombination
+from RUFAS.biophysical.animal.data_types.animal_combination import AnimalCombination
 from RUFAS.general_constants import GeneralConstants
 
 from tests.test_biophysical.test_animal.test_herd_manager.pytest_fixtures import (
