+++ tests/test_biophysical/test_manure/test_handler/test_handler.py
@@ -8,7 +8,7 @@ from RUFAS.biophysical.manure.handler.parlor_cleaning import ParlorCleaningHandl
 from RUFAS.biophysical.manure.processor import Processor
 from RUFAS.current_day_conditions import CurrentDayConditions
 from RUFAS.data_structures.animal_to_manure_connection import ManureStream, PenManureData, StreamType
-from RUFAS.enums import AnimalCombination
+from RUFAS.biophysical.animal.data_types.animal_combination import AnimalCombination
 from RUFAS.general_constants import GeneralConstants
 from RUFAS.rufas_time import RufasTime
